+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/Controller.java
@@ -6,18 +6,23 @@
  * compliance with the License.  You may obtain a copy of the
  * License at
  *
- *  http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
  * either express or implied.  See the License for the specific
  * language governing permissions and limitations under the License.
+ *
  */
 
-package org.radix.api.http;
+package com.radixdlt.api;
 
-@FunctionalInterface
-public interface ThrowingConsumer<A> {
-	void accept(A arg1) throws Exception;
+import io.undertow.server.RoutingHandler;
+
+/**
+ * A controller which can serve http requests
+ */
+public interface Controller {
+	void configureRoutes(RoutingHandler handler);
 }
