+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/JsonRpcHandler.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright 2020 Radix DLT Ltd
+ * (C) Copyright 2021 Radix DLT Ltd
  *
  * Radix DLT Ltd licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except in
@@ -13,12 +13,16 @@
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
  * either express or implied.  See the License for the specific
  * language governing permissions and limitations under the License.
+ *
  */
 
-package org.radix.api.jsonrpc;
+package com.radixdlt.api;
 
-public enum AtomStatus {
-	DOES_NOT_EXIST,
-	PENDING_CM_VERIFICATION,
-	STORED
+import org.json.JSONObject;
+
+/**
+ * Handles Json RPC requests
+ */
+public interface JsonRpcHandler {
+	JSONObject execute(JSONObject param);
 }
