+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/RestUtils.java
@@ -6,17 +6,19 @@
  * compliance with the License.  You may obtain a copy of the
  * License at
  *
- *  http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
  * either express or implied.  See the License for the specific
  * language governing permissions and limitations under the License.
+ *
  */
 
-package org.radix.api.http;
+package com.radixdlt.api;
 
+import com.radixdlt.api.archive.ThrowingConsumer;
 import org.json.JSONException;
 import org.json.JSONObject;
