+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/archive/RadixJsonRpcServer.java
@@ -1,27 +1,29 @@
 /*
- * (C) Copyright 2020 Radix DLT Ltd
+ * (C) Copyright 2021 Radix DLT Ltd
  *
  * Radix DLT Ltd licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except in
  * compliance with the License.  You may obtain a copy of the
  * License at
  *
- *  http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
  * either express or implied.  See the License for the specific
  * language governing permissions and limitations under the License.
+ *
  */
 
-package org.radix.api.jsonrpc;
+package com.radixdlt.api.archive;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.json.JSONArray;
 import org.json.JSONObject;
-import org.radix.api.jsonrpc.JsonRpcUtil.RpcError;
+import com.radixdlt.api.JsonRpcHandler;
+import com.radixdlt.api.JsonRpcUtil.RpcError;
 
 import com.google.common.io.CharStreams;
 import com.google.inject.Inject;
@@ -35,10 +37,10 @@ import java.util.Optional;
 
 import io.undertow.server.HttpServerExchange;
 
-import static org.radix.api.jsonrpc.JsonRpcUtil.errorResponse;
-import static org.radix.api.jsonrpc.JsonRpcUtil.invalidParamsError;
-import static org.radix.api.jsonrpc.JsonRpcUtil.jsonObject;
-import static org.radix.api.jsonrpc.JsonRpcUtil.parseError;
+import static com.radixdlt.api.JsonRpcUtil.errorResponse;
+import static com.radixdlt.api.JsonRpcUtil.invalidParamsError;
+import static com.radixdlt.api.JsonRpcUtil.jsonObject;
+import static com.radixdlt.api.JsonRpcUtil.parseError;
 
 /**
  * Stateless Json Rpc 2.0 Server
