+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/client/api/PreparedTransaction.java
@@ -22,7 +22,7 @@ import org.json.JSONObject;
 import com.radixdlt.utils.UInt256;
 
 import static org.bouncycastle.util.encoders.Hex.toHexString;
-import static org.radix.api.jsonrpc.JsonRpcUtil.jsonObject;
+import static com.radixdlt.api.JsonRpcUtil.jsonObject;
 
 import static java.util.Objects.requireNonNull;
