+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/client/handler/HighLevelApiHandler.java
@@ -23,7 +23,7 @@ import com.radixdlt.identifiers.REAddr;
 import com.radixdlt.crypto.HashUtils;
 import org.bouncycastle.util.encoders.Hex;
 import org.json.JSONObject;
-import org.radix.api.jsonrpc.JsonRpcUtil;
+import com.radixdlt.api.JsonRpcUtil;
 
 import com.google.common.collect.ImmutableList;
 import com.google.inject.Inject;
@@ -49,16 +49,16 @@ import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 
-import static org.radix.api.jsonrpc.JsonRpcUtil.ARRAY;
-import static org.radix.api.jsonrpc.JsonRpcUtil.fromList;
-import static org.radix.api.jsonrpc.JsonRpcUtil.invalidParamsError;
-import static org.radix.api.jsonrpc.JsonRpcUtil.jsonObject;
-import static org.radix.api.jsonrpc.JsonRpcUtil.response;
-import static org.radix.api.jsonrpc.JsonRpcUtil.safeInteger;
-import static org.radix.api.jsonrpc.JsonRpcUtil.safeString;
-import static org.radix.api.jsonrpc.JsonRpcUtil.withRequiredArrayParameter;
-import static org.radix.api.jsonrpc.JsonRpcUtil.withRequiredParameters;
-import static org.radix.api.jsonrpc.JsonRpcUtil.withRequiredStringParameter;
+import static com.radixdlt.api.JsonRpcUtil.ARRAY;
+import static com.radixdlt.api.JsonRpcUtil.fromList;
+import static com.radixdlt.api.JsonRpcUtil.invalidParamsError;
+import static com.radixdlt.api.JsonRpcUtil.jsonObject;
+import static com.radixdlt.api.JsonRpcUtil.response;
+import static com.radixdlt.api.JsonRpcUtil.safeInteger;
+import static com.radixdlt.api.JsonRpcUtil.safeString;
+import static com.radixdlt.api.JsonRpcUtil.withRequiredArrayParameter;
+import static com.radixdlt.api.JsonRpcUtil.withRequiredParameters;
+import static com.radixdlt.api.JsonRpcUtil.withRequiredStringParameter;
 
 import static com.radixdlt.utils.functional.Optionals.allOf;
 import static com.radixdlt.utils.functional.Tuple.tuple;
