+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/client/store/ActionEntry.java
@@ -31,7 +31,7 @@ import com.radixdlt.utils.UInt256;
 
 import java.util.function.Function;
 
-import static org.radix.api.jsonrpc.JsonRpcUtil.jsonObject;
+import static com.radixdlt.api.JsonRpcUtil.jsonObject;
 
 import static java.util.Objects.requireNonNull;
