+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/client/store/MessageEntry.java
@@ -21,7 +21,7 @@ import org.json.JSONObject;
 
 import java.util.Objects;
 
-import static org.radix.api.jsonrpc.JsonRpcUtil.jsonObject;
+import static com.radixdlt.api.JsonRpcUtil.jsonObject;
 
 import static java.util.Objects.requireNonNull;
